requirements.txt
numpy
pandas
plotly.express
streamlit
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4